refactor(api)!: one module per command type, and cut docs to what callers need
command_builders.rs and rithmic_command_types.rs split one type's inherent
impl by method category: validate() lived in a different file from the
build() that called it, and eleven types had impl blocks in both. The
boundary had no semantic basis — command_builders.rs was a leftover name
from the *Builder types that no longer exist. Each command now owns one
module under src/api/commands/, holding its struct, setters, validate(),
build() and tests. LoginConfig moves to config.rs; it is not a command.

The public API does not move. src/lib.rs re-exports are byte-identical and
the modules stay pub(crate), so nothing downstream can tell this happened.

OrderPlacement is renamed OrderOrigin. "Placement" reads as where or how an
order is placed; the question the enum answers is what generated it, a human
or a machine. The generated per-request enums keep the name OrderPlacement,
so the two no longer collide in prose. The field stays manual_or_auto to
match the proto.

validate() is removed from the seven command types whose body was Ok(()) —
modify, cancel, cancel-all, exit-position, bracket-level-adjustment,
link-orders and modify-order-reference-data. A public method that does
nothing is not worth documenting. build() stays on all eleven with its
signature unchanged and returns Ok(self) for those seven. validate()
survives on RithmicOrder, RithmicBracketOrder, RithmicOcoOrder and
RithmicOcoOrderLeg, and its docs now name the price-type rules rather than
describing what the method does not do.

Documentation is cut to what helps a caller integrate. The 19 compile_fail
doctests pinning #[non_exhaustive] are gone, along with the prose
introducing them: the attribute is a fact about the type, not something a
consumer needs the reasoning for. Also gone are the wire-encoding
explanations on the OCO price fields, TrailingStop's account of why it has
no Default, and the # Errors sections. Every #[non_exhaustive] attribute is
untouched — 244 on the generated types, and all the hand-written ones.

Three tests are removed, 343 to 340, each of which only asserted that a
now-deleted validate() returned Ok.
